requirements.txt
crewai
langchain-google-genai
ddgs
python-dotenv
pandas
openpyxl
xlrd
streamlit
chromadb
sentence-transformers